+++ examples/timestamp/clock_set.c
@@ -74,7 +74,7 @@ int TPM2_ClockSet_Test(void* userCtx, int argc, char *argv[])
             goto exit_badargs;
         }
         /* Otherwise we have the [time] optional argument */
-        newClock = XATOI(argv[1]);
+        newClock = (UINT64)XATOI(argv[1]);
     }
     else if (argc == 1) {
         newClock = 0;
